% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_impact.R
\name{estimate_impact}
\alias{estimate_impact}
\title{Estimate the effect size difference between focal and composite group abilities}
\usage{
estimate_impact(responses, focal_column, focal_id)
}
\arguments{
\item{responses}{The \code{data.frame} of responses, including the \code{focal_column}.}

\item{focal_column}{The \code{numeric} location of the focal column.}

\item{focal_id}{The \code{numeric}, \code{character}, or \code{logical} value that identifies
the focal group.}
}
\value{
A \code{numeric} estimate of the impact as the effect size \emph{D}, e.g., the
standardized mean theta difference between the focal group and the composite
(total) group abilities.
}
\description{
Estimate the effect size difference between focal and composite group abilities
}
