% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{timmsDiffic}
\alias{timmsDiffic}
\title{Sample TIMMS item difficulties}
\format{
An object of class \code{numeric} of length 20.
}
\usage{
timmsDiffic
}
\description{
A vector of the 20 item difficulty parameters \code{b} for the \code{timmsData} items.
}
\keyword{datasets}
