% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{timmsDiscrim}
\alias{timmsDiscrim}
\title{Sample TIMMS item discriminations}
\format{
An object of class \code{numeric} of length 20.
}
\usage{
timmsDiscrim
}
\description{
A vector of the 20 item discrimination parameters \code{a} for the \code{timmsData} items.
}
\keyword{datasets}
