% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repsd.R
\name{repsd}
\alias{repsd}
\title{repsd}
\usage{
repsd(responses = timmsData, focalColumn = 21, focalGroupID = 1, numStrata = 4)
}
\arguments{
\item{responses}{data.frame, matrix, or similar object which includes the item
responses and the focal group ID column.}

\item{focalColumn}{numeric or character. The location or name of the column
that holds the focal group data.}

\item{focalGroupID}{numeric or character. The value that identifies the focal
group.}

\item{numStrata}{numeric. How many strata for matching should be used?}
}
\value{
Matrix of repsd values for each item.
}
\description{
repsd
}
