% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repsd.R
\name{repsd}
\alias{repsd}
\title{repsd}
\usage{
repsd(
  focalSample = 88,
  focalProp = 0.9,
  responses = timmsData,
  focalColumn = 21,
  focalGroupID = 1
)
}
\arguments{
\item{focalSample}{numeric - How large is the focal sample in the data set?}

\item{focalProp}{numeric, between 0 and 1 (exclusive). What is the proportion
of the focal sample compared to the rest of the data?}

\item{responses}{data.frame, matrix, or similar object - Includes the item
responses and the focal group ID column.}

\item{focalColumn}{numeric or character - The location or name of the column
that holds the focal group data.}

\item{focalGroupID}{numeric or character - The value that identifies the focal
group.}
}
\value{
Matrix of repsd values for each item.
}
\description{
repsd
}
